refactor: extract CompilationPlugin in a diferent file
